requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via web3
aiosignal==1.3.2
    # via aiohttp
alembic==1.15.2
    # via
    #   -r requirements.in
    #   flask-migrate
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
app==0.0.1
    # via -r requirements.in
attrs==25.3.0
    # via aiohttp
billiard==4.2.1
    # via celery
bitarray==3.2.0
    # via eth-account
blinker==1.9.0
    # via
    #   flask
    #   flask-mail
build==1.2.2.post1
    # via pip-tools
celery==5.4.0
    # via -r requirements.in
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
ckzg==2.1.0
    # via eth-account
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   build
    #   click
cryptography==39.0.2
    # via -r requirements.in
cytoolz==1.0.1
    # via eth-utils
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via -r requirements.in
eth-abi==5.2.0
    # via
    #   eth-account
    #   web3
eth-account==0.13.6
    # via web3
eth-hash[pycryptodome]==0.7.1
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.6.1
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.2.0
    # via eth-account
eth-typing==5.2.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==5.2.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-login
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-script
    #   flask-sqlalchemy
flask-cors==5.0.1
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-mail==0.10.0
    # via -r requirements.in
flask-marshmallow==1.3.0
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
forex-python==1.8
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
hexbytes==1.3.0
    # via
    #   eth-account
    #   eth-rlp
    #   web3
idna==3.10
    # via
    #   email-validator
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
kombu==5.5.1
    # via celery
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.26.1
    # via
    #   -r requirements.in
    #   flask-marshmallow
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
packaging==24.2
    # via
    #   build
    #   marshmallow
parsimonious==0.10.0
    # via eth-abi
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.50
    # via click-repl
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pycryptodome==3.22.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.11.1
    # via
    #   eth-account
    #   web3
pydantic-core==2.33.0
    # via pydantic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.0
    # via -r requirements.in
pyunormalize==16.0.0
    # via web3
pywin32==310
    # via web3
regex==2024.11.6
    # via parsimonious
requests==2.32.3
    # via
    #   forex-python
    #   web3
rlp==4.1.0
    # via
    #   eth-account
    #   eth-rlp
simplejson==3.20.1
    # via forex-python
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.40
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
toolz==1.0.0
    # via cytoolz
types-requests==2.32.0.20250328
    # via web3
typing-extensions==4.13.0
    # via
    #   alembic
    #   eth-typing
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
    #   web3
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.1
    # via
    #   celery
    #   kombu
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
web3==7.10.0
    # via -r requirements.in
websockets==15.0.1
    # via web3
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   flask-login
wheel==0.45.1
    # via pip-tools
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
